+++ src/useFieldArray.ts
@@ -127,6 +127,9 @@ export function useFieldArray<
             if (Array.isArray(fieldValues)) {
               setFields(fieldValues);
               ids.current = fieldValues.map(generateId);
+            } else if (!fieldArrayName) {
+              setFields([]);
+              ids.current = [];
             }
           }
         },